// client-admin/types/order.ts
export type OrderStatus = "pending" | "confirmed" | "shipping" | "delivered" | "completed" | "cancelled" | "failed"
// ...
export interface TimelineItem {
  event: string
  time: Date
  from: OrderStatus | null
  to: OrderStatus
}